addressed review comments from @shreyas-s-rao
