+++ docs/usage/validating-etcd-clusters.md
@@ -1,7 +1,8 @@
 # CRD Validations for etcd
 
 * The validations for the fields within the `etcd` resource are done via [kubebuilder markers for CRD validation](https://book.kubebuilder.io/reference/markers/crd-validation). 
-* The validations themselves are written using a combination of [CEL expressions](https://kubernetes.io/docs/reference/using-api/cel/) via the `x-validation` tag which provides a straightforward syntax to write validation rules for the fields, and pattern matching with the use of the `validation` tag.
+* The validations for clusters with kubernetes versions `>= 1.29` are written using a combination of [CEL expressions](https://kubernetes.io/docs/reference/using-api/cel/) via the `x-validation` tag which provides a straightforward syntax to write validation rules for the fields, and pattern matching with the use of the `validation` tag.
+* The validations for clusters with kubernetes versions `< 1.29` will not contain validations via `CEL` expressions since this is GA for kubernetes version 1.29 or higher.
 * Upon any changes to the validation rules to the etcd resource, the `yaml` files for the same can be generated by running the `make generate` command.
 
 ## Validation rules:
